Cleanup: track first_line as usize.

+++ helix-term/src/editor.rs
@@ -219,7 +219,7 @@ impl Editor {
                 }
 
                 let style: Style = view.theme.get("ui.linenr");
-                for (i, line) in (view.first_line..(last_line as u16)).enumerate() {
+                for (i, line) in (view.first_line..last_line).enumerate() {
                     self.surface
                         .set_stringn(0, i as u16, format!("{:>5}", line + 1), 5, style);
                     // lavender
